requirements.txt
fastapi==0.116.1
uvicorn==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
shortuuid==1.0.13
slowapi==0.1.9
redis==6.2.0
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.28.1
psutil==7.0.0
pydantic[email]==2.11.7
google-generativeai>=0.3.2
langchain>=0.3.27
langchain-core>=0.3.72
langchain-google-genai>=2.1.8
protobuf~=4.25.8
dotenv~=0.9.9
SQLAlchemy~=2.0.41
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-cloud-translate==3.12.0
Pillow>=9.0.0